text editors - shuffle things around
